requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-cloud-aiplatform>=1.40.0
